refactor: 자동차 이름과 주어진 횟수를 검증하는 Validator 각각 분리
